feat: connected-wallet package
