+++ packages/connected-wallet/src/index.ts
@@ -1,0 +1,2 @@
+export { Near, NearConfig } from './near';
+export { ConnectedWalletAccount, WalletConnection } from './wallet-account';
